Use a DataTable or Table component depending on the number of rows
